Moved to later version of .NET and fixed a bug in LeftMostOneIndex

+++ KnuthBitops/Bitops.cs
@@ -317,7 +317,7 @@ namespace KnuthBitops
 	
 		public static int LeftmostOneIndex(int i)
 		{
-			short shifted = (short)((ushort)i >> 16);
+			short shifted = (short)((uint)i >> 16);
 			if (shifted != 0)
 			{
 				return 16 + LeftmostOneIndex(shifted);
